Converted framer into a class.
This change allows the framer to be replaced with other implementations
and still work correctly with the data link module.

Also moved fbp_framer_frame_id_subtract() from the framer to the data link module.
Using frame_id is a function of the data link layer, not the framer layer.