Changed PubSub metadata default values to be retained.
Specifying a topic default value in metadata only makes sense
if the value is also retained.  Previously, the pubsub instance
was retaining all values, so it worked.  Now that the pubsub instance
is only correctly retaining values marked with the retained flag,
the default value must also be marked as retained.